Modified tory hierarchy and removed unnessary code
In an effort to allow Tory to grow I have combined several of the
functions into the util and inventory files. These files now contain
global functions that are called from __main__.  I also updated __main__
so that it's not hard coded with the different function names.

I also redid the human readable print_nicely function so that it can
accept any python dictionary and print it out nicely.  I also renamed
several of the functions so that they follow a common naming scheme
allowing us to dynamically call them as necessary.

To Do: I Haven't implemented Kaity's database functions in yet, but it
shouldn't be too hard.